+++ examples/tokens/token_mint_transaction_non_fungible.py
@@ -9,13 +9,10 @@ python examples/token_mint_transaction_non_fungible.py
 
 import os
 import sys
-from dotenv import load_dotenv
 
 from hiero_sdk_python import (
     Client,
-    AccountId,
     PrivateKey,
-    Network,
     TokenCreateTransaction,
     TokenMintTransaction,
     TokenType,
@@ -23,26 +20,11 @@ from hiero_sdk_python import (
     ResponseCode,
 )
 
-# Load environment variables from .env file
-load_dotenv()
-network_name = os.getenv("NETWORK", "testnet").lower()
-
-
 def setup_client():
-    """Setup and return a Hedera client."""
-    network = Network(network_name)
-    print(f"Connecting to Hedera {network_name} network!")
-    client = Client(network)
-
-    try:
-        operator_id = AccountId.from_string(os.getenv("OPERATOR_ID", ""))
-        operator_key = PrivateKey.from_string(os.getenv("OPERATOR_KEY", ""))
-        client.set_operator(operator_id, operator_key)
-        print(f"Client set up with operator id {client.operator_account_id}")
-        return client, operator_id, operator_key
-    except (TypeError, ValueError):
-        print("❌ Error: Please check OPERATOR_ID and OPERATOR_KEY in your .env file.")
-        sys.exit(1)
+    client = Client.from_env()
+    print(f"Network: {client.network.network}")
+    print(f"Client set up with operator id {client.operator_account_id}")
+    return client
 
 
 def generate_supply_key():
@@ -55,7 +37,9 @@ def generate_supply_key():
 
 def create_nft_collection():
     """Create the NFT Collection (Token)"""
-    client, operator_id, operator_key = setup_client()
+
+    client = setup_client()
+
     supply_key = generate_supply_key()
     print("\nSTEP 2: Creating a new NFT collection...")
     try:
@@ -64,14 +48,14 @@ def create_nft_collection():
             .set_token_name("My Awesome NFT")
             .set_token_symbol("MANFT")
             .set_token_type(TokenType.NON_FUNGIBLE_UNIQUE)
-            .set_treasury_account_id(operator_id)
+            .set_treasury_account_id(client.operator_account_id)
             .set_initial_supply(0)  # NFTs must have an initial supply of 0
             .set_supply_key(supply_key)  # Assign the supply key for minting
         )
 
         receipt = (
             tx.freeze_with(client)
-            .sign(operator_key)
+            .sign(client.operator_private_key)
             .sign(supply_key)  # The new supply key must sign to give consent
             .execute(client)
         )