+++ examples/tokens/token_revoke_kyc_transaction.py
@@ -4,15 +4,11 @@ python examples/tokens/token_revoke_kyc_transaction.py.py
 
 """
 
-import os
 import sys
-from dotenv import load_dotenv
 
 from hiero_sdk_python import (
     Client,
-    AccountId,
     PrivateKey,
-    Network,
 )
 from hiero_sdk_python.account.account_create_transaction import AccountCreateTransaction
 from hiero_sdk_python.tokens.token_type import TokenType
@@ -28,23 +24,11 @@ from hiero_sdk_python.tokens.token_revoke_kyc_transaction import (
 )
 from hiero_sdk_python.tokens.token_create_transaction import TokenCreateTransaction
 
-load_dotenv()
-network_name = os.getenv("NETWORK", "testnet").lower()
-
-
 def setup_client():
-    """Initialize and set up the client with operator account"""
-    network = Network(network_name)
-    print(f"Connecting to Hedera {network_name} network!")
-    client = Client(network)
-
-    operator_id = AccountId.from_string(os.getenv("OPERATOR_ID", ""))
-    operator_key = PrivateKey.from_string(os.getenv("OPERATOR_KEY", ""))
-    client.set_operator(operator_id, operator_key)
+    client = Client.from_env()
+    print(f"Network: {client.network.network}")
     print(f"Client set up with operator id {client.operator_account_id}")
-
-    return client, operator_id, operator_key
-
+    return client
 
 def create_fungible_token(client, operator_id, operator_key, kyc_private_key):
     """Create a fungible token"""
@@ -159,7 +143,10 @@ def token_revoke_kyc():
     5. Granting KYC to the new account
     6. Revoking KYC from the new account
     """
-    client, operator_id, operator_key = setup_client()
+
+    client = setup_client()
+    operator_id = client.operator_account_id
+    operator_key = client.operator_private_key
 
     # Create KYC key
     kyc_private_key = PrivateKey.generate_ed25519()