+++ examples/tokens/token_update_nfts_transaction_nfts.py
@@ -4,15 +4,11 @@ python examples/tokens/token_update_transaction_nfts.py
 
 """
 
-import os
 import sys
-from dotenv import load_dotenv
 
 from hiero_sdk_python import (
     Client,
-    AccountId,
     PrivateKey,
-    Network,
 )
 from hiero_sdk_python.tokens.token_type import TokenType
 from hiero_sdk_python.response_code import ResponseCode
@@ -25,23 +21,13 @@ from hiero_sdk_python.tokens.token_update_nfts_transaction import (
 )
 from hiero_sdk_python.query.token_nft_info_query import TokenNftInfoQuery
 
-load_dotenv()
-network_name = os.getenv("NETWORK", "testnet").lower()
 
 
 def setup_client():
-    """Initialize and set up the client with operator account"""
-    network = Network(network_name)
-    print(f"Connecting to Hedera {network_name} network!")
-    client = Client(network)
-
-    operator_id = AccountId.from_string(os.getenv("OPERATOR_ID", ""))
-    operator_key = PrivateKey.from_string(os.getenv("OPERATOR_KEY", ""))
-    client.set_operator(operator_id, operator_key)
+    client = Client.from_env()
+    print(f"Network: {client.network.network}")
     print(f"Client set up with operator id {client.operator_account_id}")
-
-    return client, operator_id, operator_key
-
+    return client
 
 def create_nft(client, operator_id, operator_key, metadata_key):
     """Create a non-fungible token"""
@@ -136,7 +122,10 @@ def token_update_nfts():
     5. Updating metadata for the first NFT
     6. Verifying the updated NFT metadata
     """
-    client, operator_id, operator_key = setup_client()
+
+    client = setup_client()
+    operator_id = client.operator_account_id
+    operator_key = client.operator_private_key
 
     # Create metadata key
     metadata_private_key = PrivateKey.generate_ed25519()